+++ alembic/versions/20260424_0001_init_orion_schema.py
@@ -5,6 +5,7 @@ Revises:
 Create Date: 2026-04-24 12:00:00.000000
 
 """
+
 from typing import Sequence, Union
 
 from alembic import op